+++ lib/readwise/tag.rb
@@ -4,5 +4,9 @@ module Readwise
     :tag_id,
     :name,
     keyword_init: true
-  )
+  ) do
+    def serialize
+      to_h
+    end
+  end
 end